fixxed strlen error and others

+++ src/vga.h
@@ -36,7 +36,7 @@ static const size_t VGA_HEIGHT = 25;
 void terminal_initialize(void);
 void terminal_setcolor(uint8_t color);
 void terminal_setcursorposition(size_t x, size_t y);
-void terminal_getcursorposition(int *x, int *y);
+void terminal_getcursorposition(size_t *x, size_t *y);
 void terminal_backspace(void);
 void terminal_putentryat(char c, uint8_t color, size_t x, size_t y);
 void terminal_putchar(char c);
